#71: Move task save logic to fragment from activity
