+++ Friendly-plans/app/src/androidTest/java/pg/autyzm/friendly_plans/ExampleTest.java
@@ -1,5 +1,10 @@
 package pg.autyzm.friendly_plans;
 
+import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.assertion.ViewAssertions.matches;
+import static android.support.test.espresso.matcher.ViewMatchers.withId;
+import static android.support.test.espresso.matcher.ViewMatchers.withText;
+
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
